# Performance improvements for translations