--- a/lib/translator/filter.js
+++ b/lib/translator/filter.js
@@ -25,7 +25,8 @@ function filterByKeyPrefix(obj, keyPrefix) {
 
 /**
  * Filter translation and locales of the given language object by translation key prefix.
- * This is used by the `langJson` helper.
+ * This is used by the `langJson` helper. This method expects an object in the format
+ * of one of the language values returned by `translator/transformer.js:transform()`.
  *
  * The incoming language object looks like this:
  * {