--- a/lib/translator/filter.js
+++ b/lib/translator/filter.js
@@ -4,43 +4,72 @@
  * @module paper/lib/translator/filter
  */
 
+/*
+* Internal method to filter an object containing string keys using the given keyPrefix
+*
+* @private
+* @param {Object.<string, string>} obj
+* @param {string} keyPrefix
+* @returns {Object.<string, string>}
+*/
+function filterByKeyPrefix(obj, keyPrefix) {
+    const result = {};
+    for (const key in obj) {
+        if (typeof key === 'string' && key.startsWith(keyPrefix)) {
+            result[key] = obj[key];
+        }
+    }
+    return result;
+}
+
+
 /**
- * Filter translation and locales by translation key
+ * Filter translation and locales of the given language object by translation key prefix.
+ * This is used by the `langJson` helper.
+ *
+ * The incoming language object looks like this:
+ * {
+ *   locale: 'en',
+ *   locales: {
+ *     'salutations.welcome': 'en',
+ *     'salutations.hello': 'en',
+ *     'salutations.bye': 'en',
+ *     items: 'en',
+ *   },
+ *   translations: {
+ *     'salutations.welcome': 'Welcome',
+ *     'salutations.hello': 'Hello {name}',
+ *     'salutations.bye': 'Bye bye',
+ *     items: '{count, plural, one{1 Item} other{# Items}}',
+ *   }
+ * }
+ *
+ * The return value (assuming `keyFilter` of `salutations`, would look like this:
+ * {
+ *   locale: 'en',
+ *   locales: {
+ *     'salutations.welcome': 'en',
+ *     'salutations.hello': 'en',
+ *     'salutations.bye': 'en',
+ *   },
+ *   translations: {
+ *     'salutations.welcome': 'Welcome',
+ *     'salutations.hello': 'Hello {name}',
+ *     'salutations.bye': 'Bye bye',
+ *   }
+ * }
  * @param {Object.<string, string|Object>} language
  * @param {string} keyFilter
  * @returns {Object.<string, string|Object>}
  */
-function filterByKey(language, keyFilter) {
-    return Object.entries(language)
-        .reduce(
-            (result, [key, value]) => {
-                switch (key) {
-                    case 'translations':
-                    case 'locales':
-                        result[key] =
-                            Object.entries(value)
-                                .reduce(
-                                    (filteredInnerObject, [innerKey, innerValue]) => {
-                                        if (innerKey.startsWith(keyFilter)) {
-                                            filteredInnerObject[innerKey] = innerValue;
-                                        }
-
-                                        return filteredInnerObject;
-                                    },
-                                    {}
-                                );
-
-                        return result;
-                    default:
-                        result[key] = language[key];
-
-                        return result;
-                }
-            },
-            {}
-        );
+function filterLanguageObject(language, keyFilter) {
+    return {
+        locale: language.locale,
+        locales: filterByKeyPrefix(language.locales, keyFilter),
+        translations: filterByKeyPrefix(language.translations, keyFilter),
+    };
 }
 
 module.exports = {
-    filterByKey: filterByKey,
+    filterByKey: filterLanguageObject,
 };