--- a/lib/translator/filter.js
+++ b/lib/translator/filter.js
@@ -26,7 +26,7 @@ function filterByKeyPrefix(obj, keyPrefix) {
 /**
  * Filter translation and locales of the given language object by translation key prefix.
  * This is used by the `langJson` helper. This method expects an object in the format
- * of one of the language values returned by `translator/transformer.js:transform()`.
+ * returned by `translator/transformer.js:transform()`.
  *
  * The incoming language object looks like this:
  * {
@@ -45,7 +45,7 @@ function filterByKeyPrefix(obj, keyPrefix) {
  *   }
  * }
  *
- * The return value (assuming `keyFilter` of `salutations`, would look like this:
+ * The return value, assuming `keyFilter` of `salutations`, would look like this:
  * {
  *   locale: 'en',
  *   locales: {