--- a/lib/translator/index.js
+++ b/lib/translator/index.js
@@ -85,7 +85,6 @@ Translator.prototype.translate = function (key, parameters) {
         return this._formatFunctions[key](parameters);
     } catch (err) {
         this.logger.error(err);
-
         return '';
     }
 };
@@ -138,7 +137,6 @@ Translator.prototype._compileTemplate = function (key) {
     } catch (err) {
         if (err.name === 'SyntaxError') {
             this.logger.error(`Language File Syntax Error: ${err.message} for key "${key}"`, err.expected);
-
             return () => '';
         }