--- a/lib/translator/index.js
+++ b/lib/translator/index.js
@@ -18,27 +18,47 @@ const DEFAULT_LOCALE = 'en';
 
 /**
  * Translator constructor
+ *
  * @constructor
- * @param {string} acceptLanguage
- * @param {Object} allTranslations
+ * @param {string} acceptLanguage The Accept-Language header to be parsed to determine language to use
+ * @param {Object} allTranslations Object containing all translations coming from theme
  * @param {Object} logger
  */
 function Translator(acceptLanguage, allTranslations, logger = console) {
-    this.logger = logger;
-
-    const languages = Transformer.transform(allTranslations, DEFAULT_LOCALE, this.logger);
+    /**
+     * @private
+     * @type {Object}
+     */
+    this._logger = logger;
 
     /**
      * @private
      * @type {string}
      */
-    this._locale = LocaleParser.getPreferredLocale(acceptLanguage, languages, DEFAULT_LOCALE);
+    this._locale = LocaleParser.getPreferredLocale(acceptLanguage, Object.keys(allTranslations), DEFAULT_LOCALE);
 
     /**
      * @private
      * @type {Object.<string, string>}
+     *
+     * Looks like this:
+     * {
+     *   locale: 'en',
+     *   locales: {
+     *     'salutations.welcome': 'en',
+     *     'salutations.hello': 'en',
+     *     'salutations.bye': 'en',
+     *     items: 'en',
+     *   }
+     *   translations: {
+     *     'salutations.welcome': 'Welcome',
+     *     'salutations.hello': 'Hello {name}',
+     *     'salutations.bye': 'Bye bye',
+     *     items: '{count, plural, one{1 Item} other{# Items}}',
+     *   }
+     * }
      */
-    this._language = languages[this._locale] || {};
+    this._language = Transformer.transform(allTranslations, this._locale, DEFAULT_LOCALE, logger) || {};
 
     /**
      * @private
@@ -55,6 +75,7 @@ function Translator(acceptLanguage, allTranslations, logger = console) {
 
 /**
  * Translator factory method
+ *
  * @static
  * @param {string} acceptLanguage
  * @param {Object} allTranslations
@@ -67,6 +88,7 @@ Translator.create = function (acceptLanguage, allTranslations, logger = console)
 
 /**
  * Get translated string
+ *
  * @param {string} key
  * @param {Object} parameters
  * @returns {string}
@@ -84,13 +106,14 @@ Translator.prototype.translate = function (key, parameters) {
     try {
         return this._formatFunctions[key](parameters);
     } catch (err) {
-        this.logger.error(err);
+        this._logger.error(err);
         return '';
     }
 };
 
 /**
  * Get locale name
+ *
  * @returns {string} Translation locale
  */
 Translator.prototype.getLocale = function () {
@@ -99,6 +122,7 @@ Translator.prototype.getLocale = function () {
 
 /**
  * Get language object
+ *
  * @param {string} [keyFilter]
  * @returns {Object} Language object
  */
@@ -112,6 +136,7 @@ Translator.prototype.getLanguage = function (keyFilter) {
 
 /**
  * Get formatter
+ *
  * @private
  * @param {string} locale
  * @returns {MessageFormat} Return cached or new MessageFormat
@@ -124,6 +149,8 @@ Translator.prototype._getFormatter = function (locale) {
 
 /**
  * Compile a translation template and return a formatter function
+ *
+ * @private
  * @param {string} key
  * @return {Function}
  */
@@ -136,7 +163,7 @@ Translator.prototype._compileTemplate = function (key) {
         return formatter.compile(language.translations[key]);
     } catch (err) {
         if (err.name === 'SyntaxError') {
-            this.logger.error(`Language File Syntax Error: ${err.message} for key "${key}"`, err.expected);
+            this._logger.error(`Language File Syntax Error: ${err.message} for key "${key}"`, err.expected);
             return () => '';
         }