--- a/lib/translator/locale-parser.js
+++ b/lib/translator/locale-parser.js
@@ -7,28 +7,32 @@ const AcceptLanguageParser = require('accept-language-parser');
 const MessageFormat = require('messageformat');
 
 /**
- * Get preferred locale
- * @param {string} acceptLanguage
- * @param {Object} languages
- * @param {string} defaultLocale
+ * Parse the Accept-Language header and return the preferred locale after matching
+ * up against the list of supported locales.
+ *
+ * @param {string} acceptLanguage The Accept-Language header
+ * @param {Array} supportedLocales A list of supported locales
+ * @param {string} defaultLocale The default fallback locale
  * @returns {string}
  */
-function getPreferredLocale(acceptLanguage, languages, defaultLocale) {
-    const locale = getLocales(acceptLanguage).find(locale => languages[locale]) || defaultLocale;
+function getPreferredLocale(acceptLanguage, supportedLocales, defaultLocale) {
+    const requestedLocales = getLocales(acceptLanguage);
+    const preferredLocale = requestedLocales.find(locale => supportedLocales.includes(locale)) || defaultLocale;
 
+    // Make sure that MessageFormat supports it
     try {
-        new MessageFormat(locale);
-
-        return locale;
+        new MessageFormat(preferredLocale);
+        return preferredLocale;
     } catch (err) {
         return defaultLocale;
     }
 }
 
 /**
- * Parse locale header
- * @param {string} acceptLanguage
- * @returns {string[]} Locales
+ * Parse Accept-Language header and return a list of locales
+ *
+ * @param {string} acceptLanguage The Accept-Language header
+ * @returns {string[]} List of locale identifiers
  */
 function getLocales(acceptLanguage) {
     const localeObjects = AcceptLanguageParser.parse(acceptLanguage);