--- a/lib/translator/transformer.js
+++ b/lib/translator/transformer.js
@@ -5,8 +5,82 @@
  */
 
 /**
- * Transform translations
- * @param {Object} allTranslations
+ * Transform translations from the representation provided by stencil bundle
+ * into something that enables fast lookups in the `lang` helper.
+ *
+ * The `allTranslations` object looks like this:
+ * {
+ *   en: {
+ *     salutations: {
+ *       welcome: 'Welcome',
+ *       hello: 'Hello {name}',
+ *       bye: 'Bye bye',
+ *     }
+ *     items: '{count, plural, one{1 Item} other{# Items}}',
+ *   },
+ *   fr: {
+ *     salutations: {
+ *       hello: 'Bonjour {name}',
+ *       bye: 'au revoir',
+ *     }
+ *   },
+ *   'fr-CA': {
+ *     salutations: {
+ *       hello: 'Salut {name}',
+ *     },
+ *   },
+ * }
+ *
+ * The return value looks like this:
+ * {
+ *   en: {
+ *     locale: 'en',
+ *     locales: {
+ *       'salutations.welcome': 'en',
+ *       'salutations.hello': 'en',
+ *       'salutations.bye': 'en',
+ *       items: 'en',
+ *     },
+ *     translations: {
+ *       'salutations.welcome': 'Welcome',
+ *       'salutations.hello': 'Hello {name}',
+ *       'salutations.bye': 'Bye bye',
+ *       items: '{count, plural, one{1 Item} other{# Items}}',
+ *     }
+ *   },
+ *   fr: {
+ *     locale: 'fr',
+ *     locales: {
+ *       'salutations.welcome': 'en',
+ *       'salutations.hello': 'fr',
+ *       'salutations.bye': 'fr',
+ *       items: 'en',
+ *     },
+ *     translations: {
+ *       'salutations.welcome': 'Welcome',
+ *       'salutations.hello': 'Bonjour {name}',
+ *       'salutations.bye': 'au revoir',
+ *       items: '{count, plural, one{1 Item} other{# Items}}',
+ *     }
+ *   },
+ *   'fr-CA': {
+ *     locale: 'fr-CA',
+ *     locales: {
+ *       'salutations.welcome': 'en',
+ *       'salutations.hello': 'fr-CA',
+ *       'salutations.bye': 'fr',
+ *       items: 'en',
+ *     },
+ *     translations: {
+ *       'salutations.welcome': 'Welcome',
+ *       'salutations.hello': 'Salut {name}',
+ *       'salutations.bye': 'au revoir',
+ *       items: '{count, plural, one{1 Item} other{# Items}}',
+ *     }
+ *   },
+ * }
+ *
+ * @param {Object.<string, Object>} allTranslations
  * @param {string} defaultLocale
  * @param {Object} logger
  * @returns {Object.<string, Object>} Transformed translations
@@ -16,10 +90,54 @@ function transform(allTranslations, defaultLocale, logger = console) {
 }
 
 /**
- * Flatten translations
- * @param {Object} allTranslations
+ * Flatten translation keys to a single top-level namespace.
+ *
+ * The `allTranslations` object looks like this:
+ * {
+ *   en: {
+ *     salutations: {
+ *       welcome: 'Welcome',
+ *       hello: 'Hello {name}',
+ *       bye: 'Bye bye',
+ *       formal: {
+ *           bye: 'Farewell',
+ *       }
+ *     }
+ *     items: '{count, plural, one{1 Item} other{# Items}}',
+ *   },
+ *   fr: {
+ *     salutations: {
+ *       hello: 'Bonjour {name}',
+ *       bye: 'au revoir',
+ *     }
+ *   },
+ *   'fr-CA': {
+ *     salutations: {
+ *       hello: 'Salut {name}',
+ *     },
+ *   },
+ * }
+ *
+ * The return value looks like this:
+ * {
+ *   en: {
+ *     'salutations.welcome': 'Welcome',
+ *     'salutations.hello': 'Hello {name}',
+ *     'salutations.bye': 'Bye bye',
+ *     'salutations.formal.bye': 'Farewell',
+ *     items: '{count, plural, one{1 Item} other{# Items}}',
+ *   },
+ *   fr: {
+ *     'salutations.hello': 'Bonjour {name}',
+ *     'salutations.bye': 'au revoir',
+ *   },
+ *   'fr-CA': {
+ *     'salutations.hello': 'Salut {name}',
+ *   },
+ * }
+ * @param {Object.<string, Object>} allTranslations
  * @param {Object} logger
- * @returns {Object.<string, Object>} Flatten translations
+ * @returns {Object.<string, Object>} Flattened translations
  */
 function flatten(allTranslations, logger = console) {
     return Object.entries(allTranslations)
@@ -29,7 +147,6 @@ function flatten(allTranslations, logger = console) {
                     result[locale] = flattenObject(translation);
                 } catch (err) {
                     logger.error(`Failed to parse ${locale} - Error: ${err}`);
-
                     result[locale] = {};
                 }
 
@@ -40,13 +157,82 @@ function flatten(allTranslations, logger = console) {
 }
 
 /**
- * Cascade translations
- * @param {Object} allTranslations Flattened translations
+ * Cascade translations by providing appropriate fallback values. For example, if 'fr-CA'
+ * is requested but the translation override doesn't exist, we fallback first to 'fr', then
+ * to the defaultLocale (usually 'en').
+ *
+ * flattenedTranslations looks like this:
+ * {
+ *   en: {
+ *     'salutations.welcome': 'Welcome',
+ *     'salutations.hello': 'Hello {name}',
+ *     'salutations.bye': 'Bye bye',
+ *     items: '{count, plural, one{1 Item} other{# Items}}',
+ *   },
+ *   fr: {
+ *     'salutations.hello': 'Bonjour {name}',
+ *     'salutations.bye': 'au revoir',
+ *   },
+ *   'fr-CA': {
+ *     'salutations.hello': 'Salut {name}',
+ *   },
+ * }
+ *
+ * The return value looks like this:
+ * {
+ *   en: {
+ *     locale: 'en',
+ *     locales: {
+ *       'salutations.welcome': 'en',
+ *       'salutations.hello': 'en',
+ *       'salutations.bye': 'en',
+ *       items: 'en',
+ *     },
+ *     translations: {
+ *       'salutations.welcome': 'Welcome',
+ *       'salutations.hello': 'Hello {name}',
+ *       'salutations.bye': 'Bye bye',
+ *       items: '{count, plural, one{1 Item} other{# Items}}',
+ *     }
+ *   },
+ *   fr: {
+ *     locale: 'fr',
+ *     locales: {
+ *       'salutations.welcome': 'en',
+ *       'salutations.hello': 'fr',
+ *       'salutations.bye': 'fr',
+ *       items: 'en',
+ *     },
+ *     translations: {
+ *       'salutations.welcome': 'Welcome',
+ *       'salutations.hello': 'Bonjour {name}',
+ *       'salutations.bye': 'au revoir',
+ *       items: '{count, plural, one{1 Item} other{# Items}}',
+ *     }
+ *   },
+ *   'fr-CA': {
+ *     locale: 'fr-CA',
+ *     locales: {
+ *       'salutations.welcome': 'en',
+ *       'salutations.hello': 'fr-CA',
+ *       'salutations.bye': 'fr',
+ *       items: 'en',
+ *     },
+ *     translations: {
+ *       'salutations.welcome': 'Welcome',
+ *       'salutations.hello': 'Salut {name}',
+ *       'salutations.bye': 'au revoir',
+ *       items: '{count, plural, one{1 Item} other{# Items}}',
+ *     }
+ *   },
+ * }
+ *
+ * @param {Object.<string, Object>} flattenedTranslations Flattened translations
  * @param {string} defaultLocale
- * @returns {Object.<string, Object>} Language objects
+ * @returns {Object.<string, Object>} Cascaded translations spec
  */
-function cascade(allTranslations, defaultLocale) {
-    return Object.entries(allTranslations)
+function cascade(flattenedTranslations, defaultLocale) {
+    return Object.entries(flattenedTranslations)
         .reduce(
             (result, [locale, translations]) => {
                 if (!result[locale]) {
@@ -56,7 +242,7 @@ function cascade(allTranslations, defaultLocale) {
                 const regionCodes = locale.split('-');
                 for (let regionIndex = regionCodes.length - 1; regionIndex >= 0; regionIndex--) {
                     const parentLocale = getParentLocale(regionCodes, regionIndex, defaultLocale);
-                    const parentTranslations = allTranslations[parentLocale] || {};
+                    const parentTranslations = flattenedTranslations[parentLocale] || {};
 
                     new Set(
                         Object.keys(parentTranslations).concat(Object.keys(translations))
@@ -94,7 +280,9 @@ function getParentLocale(regionCodes, regionIndex, defaultLocale) {
 }
 
 /**
- * Brings nested JSON parameters to the top level while preserving namespaces with . as a delimiter
+ * Internal method to flatten nested JSON to the top level, tranforming nested keys
+ * using dot syntax.
+ *
  * @private
  * @param {Object} object
  * @param {Object} [result={}]