… spec/fixtures/lang.json
… spec/lib/translator.js