* Improve performance of filterByKey by using imperative style
* Improve performance of Translator constructor - use imperative style, and only process data for the preferred language
* Note that this PR updates the interface for the three functions in `lib/translator/transformer.js` as well as `getPreferredLocale` in `lib/translator/locale-parser.js`, but these methods should be considered private to the `Translator` class so should not constitute a breaking change. The interface for `Translator` remains the same.
* Add docs around data structures used in translations
* Add test coverage for transform, flatten, cascade

## filterByKey


#### Before
```
[matt.olson@C02X425EJG5M:~/workspace/paper] $ ./node_modules/lab/bin/lab -v spec/lib/filter.js
Filter
  ✔ 1) should return locale unchanged (1 ms)
  ✔ 2) should return filtered locales (0 ms)
  ✔ 3) should return filtered translations (1 ms)
  ✔ 4) load test (3607 ms)


4 tests complete
Test duration: 3631 ms
```

#### After
```
[matt.olson@C02X425EJG5M:~/workspace/paper] $ ./node_modules/lab/bin/lab -v spec/lib/filter.js
Filter
  ✔ 1) should return locale unchanged (1 ms)
  ✔ 2) should return filtered locales (1 ms)
  ✔ 3) should return filtered translations (0 ms)
  ✔ 4) load test (1296 ms)


4 tests complete
Test duration: 1319 ms
```

## Translator constructor
#### Before
```
[matt.olson@C02X425EJG5M:~/workspace/paper] $ ./node_modules/lab/bin/lab -v spec/lib/translator.js
Translator
  ✔ 1) should return translated strings (5 ms)
  ✔ 2) should return translated strings in default language if it cannot find a translation file for a specified language (1 ms)
  ✔ 3) should return translated strings in English if cannot locate the preferred translation file (1 ms)
  ✔ 4) should return translated strings in English if the preferred locale name is invalid (1 ms)
  ✔ 5) should return translated strings in English if specific keys are missing from the preferred translation file (0 ms)
  ✔ 6) should return translated strings in English and print to log if the translation file cannot be parsed (4 ms)
  ✔ 7) should return translated strings in the most region-specific language if possible (1 ms)
  ✔ 8) should return an empty string and log a message if missing required parameters (1 ms)
  ✔ 9) should log an error when there is a syntax error in the template (1 ms)
  ✔ 10) should log an error when there is a syntax error in the template (1 ms)
  ✔ 11) should return the translation key if both the preferred and fallback translations are missing (1 ms)
  ✔ 12) should return pluralized strings according to their language (1 ms)
  ✔ 13) should return the current locale name (0 ms)
  ✔ 14) should return a translation object (1 ms)
  ✔ 15) should return a cascaded translation object (0 ms)
  ✔ 16) should return a translation object filtered by key (1 ms)
  ✔ 17) load test (118692 ms)


17 tests complete
Test duration: 118720 ms
```

#### After
```
[matt.olson@C02X425EJG5M:~/workspace/paper] $ ./node_modules/lab/bin/lab -v spec/lib/translator.js
Translator
  ✔ 1) should return translated strings (4 ms)
  ✔ 2) should return translated strings in default language if it cannot find a translation file for a specified language (1 ms)
  ✔ 3) should return translated strings in English if cannot locate the preferred translation file (1 ms)
  ✔ 4) should return translated strings in English if the preferred locale name is invalid (1 ms)
  ✔ 5) should return translated strings in English if specific keys are missing from the preferred translation file (1 ms)
  ✔ 6) should return translated strings in English and fail gracefully if the translation file cannot be parsed (5 ms)
  ✔ 7) should return translated strings in the most region-specific language if possible (0 ms)
  ✔ 8) should return an empty string and log a message if missing required parameters (0 ms)
  ✔ 9) should log an error when there is a syntax error in the template (0 ms)
  ✔ 10) should log an error when there is a syntax error in the template (1 ms)
  ✔ 11) should return the translation key if both the preferred and fallback translations are missing (0 ms)
  ✔ 12) should return pluralized strings according to their language (1 ms)
  ✔ 13) should return the current locale name (0 ms)
  ✔ 14) should return a translation object (0 ms)
  ✔ 15) should return a cascaded translation object (0 ms)
  ✔ 16) should return a translation object filtered by key (1 ms)
  ✔ 17) load test (6106 ms)


17 tests complete
Test duration: 6135 ms
```